Fix JS crash "ReferenceError: babelHelpers is not defined"
abortcontroller-polyfill is using rollup + babel to transpile
modern JS into ES5 checked into dist/, so that other projects
can import these directly regardless whether these projects
are using babel themselves or not.

This crash was happening when Babel was used in a project
that in turn was using abortcontroller-polyfill, and when
the abortcontroller-polyfill dependency was overridden using
"npm link". In that case, the external-helpers declaration
in the abortcontroller-polyfill would be picked up by babel
running in the parent project causing the "missing babelHelpers"
JS crash.

"External Helpers" are generic Babel helper functions that
you can choose to not emit in the beginning of every file
if you prefer to provide them via other means, more here:
https://babeljs.io/docs/plugins/external-helpers/

This github issue was has info on a similar issue:
https://github.com/storybooks/storybook/issues/1320#issuecomment-310777396